refactor(ui): split KeymapEditor.tsx into 11 focused modules (2446→479 lines)
Extract 11 files from the monolithic KeymapEditor component to bring all
files under their category size limits:

- keymap-editor-types.ts: shared types, constants, and interfaces
- keymap-editor-toolbar.tsx: IconTooltip, ScaleInput, toggleButtonClass
- KeyboardPane.tsx: keyboard pane display component
- LayerListPanel.tsx: layer list panel with inline rename
- HistoryToggle.tsx: typing test history modal toggle
- QmkSettingsModal.tsx: data-driven QMK settings modal renderer
- TypingTestPane.tsx: typing test UI section
- useInputModes.ts: matrix tester + typing test lifecycle hook
- useKeymapMultiSelect.ts: multi-selection state management hook
- useKeymapSelectionHandlers.ts: selection, popover, undo, keycode handlers
- useLayoutOptionsPanel.ts: layout options, overlay, min-height calc

KeymapEditor.tsx is now a pure orchestrator wiring hooks and components.
All 11 test files continue to import from KeymapEditor.tsx unchanged.
